feat: Support Cypress 10
BREAKING CHANGE: In Cypress 10, the 'integrationFolder' was removed from the
config. This value had been used to determine where to save and load the request
fixtures, so a new approach has been implemented and the plugin cannot find
previously created fixtures.

BREAKING CHANGE: The file extension of the recorded fixtures was changed to
'.cy-playback' to make it easier to identify that these files belong to the
plugin.